Little fixes

+++ classparser-bytecode/src/main/java/com/classparser/bytecode/impl/assembly/AgentAssembler.java
@@ -5,6 +5,7 @@ import com.classparser.bytecode.impl.agent.Agent;
 import com.classparser.bytecode.impl.agent.RetransformClassStorage;
 import com.classparser.bytecode.impl.assembly.attach.AgentAttacher;
 import com.classparser.bytecode.impl.assembly.build.AgentBuilder;
+import com.classparser.bytecode.impl.configuration.ConfigurationManager;
 
 import java.io.File;
 
@@ -16,26 +17,57 @@ public class AgentAssembler {
 
     private final AgentAttacher agentAttacher;
 
+    private ConfigurationManager configurationManager;
+
     public AgentAssembler(AgentAttacher agentAttacher) {
         this.agentAttacher = agentAttacher;
     }
 
+    public void setConfigurationManager(ConfigurationManager configurationManager) {
+        this.configurationManager = configurationManager;
+    }
+
     public void assembly(JavaAgent agent) {
         if (!agent.isInitialize()) {
-            String agentPath = AgentBuilder.getBuilder()
-                    .addAgentName(getAgentJarName())
-                    .addAgentClass(getAgentClass())
-                    .addManifest(getManifestFileName())
-                    .addClasses(getAgentJarClasses())
-                    .build();
-
+            String agentPath;
+            if (isCacheAgentJar()) {
+                String preparedAgentPath = AgentBuilder.getAgentPath(getAgentJarName());
+                File file = new File(preparedAgentPath);
+                if (!file.exists()) {
+                    agentPath = createAgent();
+                } else {
+                    agentPath = preparedAgentPath;
+                }
+            } else {
+                agentPath = createAgent();
+            }
 
             agentAttacher.attach(agentPath);
-            File agentJar = new File(agentPath);
-            if (!agentJar.exists() || !agentJar.delete()) {
-                System.err.println("Problems occurred with removing agent jar file: " + getAgentJarName());
+
+            if (!isCacheAgentJar()) {
+                File agentJar = new File(agentPath);
+                if (!agentJar.exists() || !agentJar.delete()) {
+                    System.err.println("Problems occurred with removing agent jar file: " + getAgentJarName());
+                }
             }
         }
+    }
+
+    private boolean isCacheAgentJar() {
+        if (configurationManager != null) {
+            return configurationManager.isCacheAgentJar();
+        }
+
+        return false;
+    }
+
+    private String createAgent() {
+        return AgentBuilder.getBuilder()
+                .addAgentName(getAgentJarName())
+                .addAgentClass(getAgentClass())
+                .addManifest(getManifestFileName())
+                .addClasses(getAgentJarClasses())
+                .build();
     }
 
     protected String getAgentJarName() {
